+++ src/mjlab/asset_zoo/robots/agibot_x2/__init__.py
@@ -1,0 +1,1 @@
+"""AgiBot X2 Ultra humanoid."""